refactor(comp/persist): migrate core financial entities to BigDecimal
Migrates VirtualPortfolioPosition, VirtualPortfolioTrade, and TickData to
use BigDecimal for money values. Updates computation engines, filters,
and repositories to use BigDecimal arithmetic and appropriate SQL column
reads. Retains double for dimensionless technical indicators and rates.

+++ computation/src/main/java/com/tickonomics/computation/demo/FillEstimate.java
@@ -1,8 +1,10 @@
 package com.tickonomics.computation.demo;
 
+import java.math.BigDecimal;
+
 /**
  * Simulated fill produced by an execution handler ({@link PassiveExecutionHandler} or
  * {@link SniperExecutionHandler}); converted into {@code ComparativeExecutionAnalysis.ExecutionSlippage}
  * samples so passive vs. aggressive strategies can be compared in the demo report.
  */
-public record FillEstimate(String type, double fillPrice, double slippageBps, boolean filled) {}
+public record FillEstimate(String type, BigDecimal fillPrice, double slippageBps, boolean filled) {}
